Backend/pyproject.toml
[project]
name = "starfunc-server"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "gunicorn>=22.0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",

    # Redis
    "redis[hiredis]>=5.0",

    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",

    # Validation & Settings
    "pydantic>=2.7",
    "pydantic-settings>=2.2",

    # Logging & Monitoring
    "structlog>=24.1",
    "prometheus-fastapi-instrumentator>=7.0",
    "prometheus-client>=0.20",

    # HTTP client (for external auth verification)
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "testcontainers[postgres,redis]>=4.4",
    "fakeredis>=2.20",
    "factory-boy>=3.3",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
